+++ backend/src/main/java/com/example/fifotech/repository/JobPostingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobPostingRepository extends JpaRepository<JobPosting, Integer> {
+public interface JobPostingRepository extends JpaRepository<JobPosting, Long> {
